(PC-14143)[PRO] refactor: SirenField: fix import
