+++ pro/src/components/layout/form/fields/SirenField/index.ts
@@ -1,0 +1,1 @@
+export { default as SirenField } from './SirenField'
