+++ pro/src/components/pages/Signup/SignupForm/SignupForm.jsx
@@ -6,7 +6,7 @@ import { Link } from 'react-router-dom'
 
 import FieldErrors from 'components/layout/form/FieldErrors'
 import PasswordField from 'components/layout/form/fields/PasswordField'
-import SirenField from 'components/layout/form/fields/SirenField/SirenField'
+import { SirenField } from 'components/layout/form/fields/SirenField'
 import TextInput from 'components/layout/inputs/TextInput/TextInput'
 import LegalInfos from 'components/layout/LegalInfos/LegalInfos'
 import { redirectLoggedUser } from 'components/router/helpers'
